Remove S3 Type and change Swift variable names
We no longer use a specific variable to determine what kind of S3 we
want to use. Instead, the code will see if we have specified a Swift
Endpoint, in which case, we use Swift.

We now use separate variables for Swift credentials.

Added relevant config to the PropertyDefinitionListBuilder of each
module.